+++ include/terrain.hpp
@@ -31,7 +31,7 @@ public:
   unsigned char fetchHeight(glm::ivec2 v) const;
   double getHeight(double x, double y) const;
 
-  void screenResized(const glm::mat4& projMat, GLuint, GLuint) override;
+  void screenResized(const glm::mat4& projMat, size_t, size_t) override;
   void render(float time, const engine::Camera& cam) override;
 };
